requirements.txt
playwright==1.41.0
playwright-stealth==1.0.6
requests
python-dotenv
setuptools
